requirements.txt
Flask==2.0.2
py_cui==0.1.4
PyNaCl==1.4.0
requests==2.26.0
peewee==3.14.8
python-dateutil==2.8.2